Limit memory of IDG buffer to 16 GB max.
Fixes #2. On big-mem machines with 512 GB, WSClean would claim around 200 GB just for the buffer, even if the MS is only 1 GB itself.
Even just claiming the mem takes a lot of time, so this not only is a nasty habbit for destroying cache/avail mem, it only makes
WSClean slow. Bigger buffers make IDG faster, but they don't need to be bigger than about the mem on the GPU. 16 GB seems to be a good
balance.