Target CPU selection
This feature provides a way to select a target CPU for compilation during build (using `TARGET_CPU=<cpu>`), unless the user explicitly requested a `PORTABLE` build. By default, sources will be compiled for the native CPU, which will then (obviously) become the target CPU. The name of the target CPU (if any) will become part of the Debian package name when building packages (i.e. `BUILD_PACKAGES=ON`). The package will also contain a the list of conflicting packages, due to the fact that binaries compiled for different target CPUs may conflict with one another.