Change least-squares fitting to maximize power on integrated image
E. Ceccotti found that cleaning might get stuck in a loop. Because the
least squares fit finds the inverse-variance-optimal fit, it does not
have to be the same as what the integrated image points out as the
strongest component. This could lead to getting stuck on a component
that has both negative and positive values:
max component is 1 Jy, but the least-squares fit gives due to
'different weighting' of the channels a flux of 0 Jy, and nothing
is removed from the image, so the max component remains 1 Jy.