AST-629 Refactor MSRowProvider
In order to create a BdaMsRowProvider it's required to use a class in the
same inheritance tree as MSRowProvider. This moves the minimal interface
to a new base class. During the implementation of AST-630 some minor
changes may still be required.

The PartitionedMS now uses the MsRowProviderBase instead of MSRowProvider
as its row provider. In prepation it uses a factory for the
DirectMSRowProvider, this factory can easily be extended to create a
BdaMsRowProvider when the measurement set has BDA data.