PAN-337 Optimise ApplyConjugatedParmResponse when nparms==2
When nparms == 2 AverageCorrection::Add is called with a matrix that has zero diagonals. A large portion of the computation done on this matrix is unnecessarry because of the 16 results in the final KroneckerProduct 12 are guaranteed to be zero and don't need to be computed. Also various transpose operations are called that only swap the two zeros around. The compiler is unable to optimize this away. This merge request adds a new AverageCorrection::AddDiagonals that performs an optimised version of the KroneckerProduct that takes this into account, allowing the optimising compiler to do a far better job of optimization.

Also make some other minor diagonal matrix optimisations to the visibility modifier code and clean it up a bit.
Also remove some duplicated map lookups that were not being optimised away (and the code is anyway cleaner with them removed)