PAN-315 Introduce overlapping IO for msgriddermanager
When `shared-facet-reads` is active; By reading the next chunk of data to be gridded while the first is busy processing we can get a sizeable performance improvement due to less core idle time. This is because the gridding parallises well to consume all cores while the reading/pre-processing of the data is mostly serial. This commit implements an overlapping IO scheme that exploits the above.

The code is also broken out into various functions to make it a bit cleaner and easier to follow.

Gives around a 25% performance improvement to gridding when running with `-shared-facet-reads` in some circumstances.