MERL-65 Remove duplicate/redundant writes of model images when using facets
The exact same duplicate model image was written for each facet in a "facet group" instead of once per group. This required both a read and write of potentially substantial amounts of data per image which was expensive both in terms of IO and saturating the network bandwidth of the cluster.

Change this so that the model image is only written once per "facet group".

On large multi node runs with specific settings these redundant writes were accounting for around 20% of program runtime in benchmarks, with this change showing a 1.2x speedup in those cases.