MERL-94 Fix shared-read behavior for direction dependent PSFs
When running with `-shared-facet-reads` and also enabling direction dependent PSFs(`-dd-psf-grid`) the generated PSFs were invalid.

This was because of the rotation applied to the PSF values, these were being applied once at the start of gridding, but with a shared buffer for all "facets" this is not possible to do in a valid way, instead rotations need to be applied "on the fly" along with other changes we already apply "on the fly" (e.g. corrections).

This MR alters the way rotations are applied so that `-shared-facet-reads` and `-dd-psf-grid` can be used together correctly.

Some additional code improvements are done to make these changes more manageable:
* `CalculateWeights()` is split into `ComputeVisibilities()` and `ComputeVisibilities()`
* Remaining parameters of these methods are made const correct which makes it easier to understand what their intention is.
* `RotateVisibilities()` is split into several methods so that it can be used in both places without code duplication.
* Some dead code is removed and other minor code cleanup (variable renames etc.)