MERL-71 Allow overlapped processing of two predict tasks
This enables two related optimizations:

1. IO overlap; second predict chunk can start processing while first is writing.
2. Processing overlap; second predict chunk can start processing its first facets while second is processing its last facets.

These changes reduce the amount of time that all/some cores are idle waiting for work and improves throughput as a result. My benchmark setup is showing a 1.25x performance improvement from these changes (for predict not whole program)

Optimization (1) is somewhat obvious but (2) deserves a slightly more thorough explanation:

* If we have 36 cores and we are processing 6 facets at a time in parallel, 6 cores per facet.
* If we have processed 28 of 32 facets with the last 4 still processing.
* Then at this point in time we have 2\*6 = 12 idle cores.
* If another 2 facets finish then we have 4\*6 = 24 idle cores.
* If another 1 finishes then we temporarily have 5\*6 = 30 idle cores.
* As we near the end of processing a chunk we have more and more idle cores.

With two overlapping chunks the initial tasks from the second processing chunk will make use of these cores and this will reduce (or potentially eliminate) the amount and time that cores are idle.