MERL-98 Use chunked mutex for ExpandAndCombineFacetVisibilities instead of single mutex to reduce contention
Alter `ExpandAndCombineFacetVisibilities()` to sum the combined visibilities in smaller ranges/chunks, with a mutex per chunk, instead of with one single mutex.

This helps reduce the time spent waiting for a lock, while not making the locking so fine grained that contention becomes high.

Benchmarks on a 5 node 96 core run are showing a `1.31x` speedup for `-shared-facet-writes`; this may be even higher in some situations (e.g. a node with more cores, or different IO speeds etc.)

A different "lock free" approach using atomic add operations and `memory_order_relaxed` was trialed but achieved only a `1.15x` speedup so the chunk/ranged option was picked instead.