MERL-103 Initial fix for beams with shared-facet-reads
Entire beam was not being cached for each facet, therefore when applying the beam during the "on the fly" gridding callback the correct beam was not applied. As the amount of data is quite large we have to be careful ab out how this is handled, we can't just cache for the entire MS.

* A new struct `BeamResponseCacheChunk` is introduced to help manage the caching.
* For each chunk processed an instance of `BeamResponseCacheChunk` is created.
* `StartProcessingChunk()` needs to be called before each new chunk processed.
* `FinishProcessingChunk()` needs to be called after each new chunk processed.
* `FinishChunkedProcessing()` needs to be called when we finish reading the data for the final chunk.
* When processing a chunk `TakeCachedBeamResponse()` needs to be used to take ownership of the chunk.
* Smart pointers are used to track ownership so that chunks can immediately be freed once no longer needed.
* `CacheBeamResponse` is updated to populate beam response chunks when the entire beam response is requested and split into two new helpers that it calls for this functionality:
  * `UpdateCachedBeamResponseForRow `- Used when caching only the most recent response (old/current behavior)
  * `CacheBeamResponseWithCacheChunk` - Used when caching the entire beam response in chunks (new shared-read behavior)

Fix some race condition issues in shared read/write code that were causing intermittent CI failures.

Also remove `SortFacetTasks()` optimization, it seems to be causing problems, also other changes made to the code since this were added may mean that the optimization is no longer really needed. This optimization can in future be revisited, tested more thoroughly and redone if it is shown to still be worthwhile.