MERL-68 Only reset model when necessarry and use sparse file if possible
With -shared-facet-reads its not always necessary to add-assign when writing to the model file; which in turn means its not always necessary to have a zero initialized model file.
As a performance optimization we can take advantage of this and only reset the model file when absolutely necessarry.

Additionally as a further optimization (that applies also for not using shared reads) we can make use of sparse files (for reordered file writers) in cases where it is necesarry to do the reset, instead of taking lots of time to write the entire file of zeros to disk, similar to how we are now using sparse files to create the initial file.

This MR is a reduced part of a larger MR, that also introduced sparse files for initial model creation for shared reads. The larger MR became conflicting/redundant when parts of it were adapted into a different MR that also makes use of sparse files when not doing shared reads. This MR re-implements remaining parts of the MR that still needed to be done.
Below is text from the original MR for context as it still explains why these changes were ultimately necessarry:
"Multi node benchmarks have shown that with enough nodes, writing and resetting the model file starts to become a significant portion of overall runtime. As a serial bottleneck, even though less expensive than e.g. inversion/prediction on a single node this becomes an obstacle to scaling. A 1.22x performance improvement was measured with very specific multi node test parameters. Note that this does not mean that such an improvement can be expected in all cases."
The original MR can be found here:
https://gitlab.com/aroffringa/wsclean/-/merge_requests/856